requirements.txt
python-dotenv
typing_extensions
anthropic